Render Character

+++ src/main/java/structures/Vector2D.java
@@ -1,5 +1,5 @@
 package structures;
 
-public record Vector2D(float x, float y) {
+public record Vector2D(int x, int y) {
 
 }
